fixed bug: reading recvwindow

+++ client.h
@@ -9,7 +9,7 @@ struct Config {
     char serverAddr[MAXCHAR];
     int port;
     char dataFile[MAXCHAR];
-    short int recvWinSize;
+    unsigned int recvWinSize;
     int seed;
     float pLoss;
     unsigned int mu;
